# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Saudi Bookkeeping Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments/allocations/{id}/unallocate:
    post:
      operationId: unallocate
      tags: [payments]
      summary: "Phase A — correct an allocation with a superseding record: Dr AR / Cr Customer deposits (or credit balances); the original allocation stays visible and untouched"
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UnallocateInput" }
      responses:
        "200":
          description: The allocation, now reversed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaymentAllocationDetail" }
        "400":
          description: Missing reason
        "404":
          description: Not found
        "409":
          description: "`allocation_already_corrected` (one correction per allocation — concurrent attempts get this); `credit_note_original_application_immutable`"
        "422":
          description: "`customer_required_for_unapplied`"
        "423":
          description: "`period_closed` — corrections are dated today; a closed current month refuses"
...
components:
  schemas:
...
    AllocationReversal:
      type: object
      required: [id, journalEntryId, reason, createdAt]
      properties:
        id: { type: integer }
        journalEntryId: { type: integer, description: "The correcting entry (UNALLOC-<id>): Dr AR / Cr the origin's liability." }
        reason: { type: string }
        createdAt: { type: string }
    PaymentAllocation:
      type: object
      required: [id, invoiceId, amount, journalEntryId, createdAt, reversedBy]
      properties:
        id: { type: integer }
        invoiceId: { type: integer }
        amount: { type: number }
        journalEntryId: { type: ["integer", "null"], description: "Set when this allocation posted its own entry (a later allocation of a deposit); null when folded into the receipt's entry." }
        createdAt: { type: string }
        reversedBy:
          oneOf:
            - { $ref: "#/components/schemas/AllocationReversal" }
            - { type: "null" }
          description: "Phase A — the correction that superseded this allocation, or null while it is active. The allocation row itself is never edited."
    PaymentAllocationDetail:
      allOf:
        - { $ref: "#/components/schemas/PaymentAllocation" }
        - type: object
          required: [paymentId, creditNoteId]
          properties:
            paymentId: { type: ["integer", "null"] }
            creditNoteId: { type: ["integer", "null"] }
    UnallocateInput:
      type: object
      required: [reason]
      properties:
        reason: { type: string, minLength: 1, maxLength: 500 }
        idempotencyKey: { type: ["string", "null"], maxLength: 120 }
...